+++ packages/terra-graphs-docs/src/example-datasets/graphConfigObjects/General/generalDefaultYAxisHidden.js
@@ -1,0 +1,32 @@
+const getDefaultConfig = (id) => ({
+  bindTo: id,
+  bindLegendTo: null,
+  axis: {
+    x: {
+      show: true,
+      label: 'Data',
+      lowerLimit: 80,
+      upperLimit: 280,
+    },
+    y: {
+      show: false,
+      label: 'Line Set A',
+      lowerLimit: -18,
+      upperLimit: 18,
+    },
+    y2: {
+      show: false,
+      label: 'Line Set B',
+      lowerLimit: 0,
+      upperLimit: 250,
+    },
+  },
+  showLabel: true,
+  showLegend: true,
+  showShapes: true,
+  showVGrid: true,
+  showHGrid: true,
+});
+
+export default getDefaultConfig;
+