+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/BarGraph/BasicBarGraphs/NegativeBars.jsx
@@ -4,27 +4,23 @@ import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Graph.module.scss';
 import '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Bar/BarGraph.module.scss';
 import ExampleGraphContainer from '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer';
-import barDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bar/barDefault';
-import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/negativeBarData';
+import getBarConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bar/barDefaultNegativeAxis';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/dataset1negative';
 
 /*
- Please refer to the documentation below to see the graphConfig and data objects
- */
+Please refer to the documentation below to see the graphConfig and data objects
+*/
 
-const graphConfig = utils.deepClone(barDefault('#negativeBargraph'));
-graphConfig.axis.y.lowerLimit = -15;
-graphConfig.axis.y.upperLimit = 0;
+const graphConfig = utils.deepClone(getBarConfig('#negativeBarGraph'));
 const dataset = utils.deepClone(exampleData);
 
 export default () => {
   React.useEffect(() => {
     const graph = Carbon.api.graph(graphConfig);
-    dataset.forEach((data) => {
-      graph.loadContent(Carbon.api.bar(data));
-    });
+    graph.loadContent(Carbon.api.bar(dataset));
   }, []);
 
   return (
-    <ExampleGraphContainer id="negativeBargraph" />
+    <ExampleGraphContainer id="negativeBarGraph" />
   );
 };