+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/BarGraph/BasicBarGraphs/AxisInfoTextLabels.jsx
@@ -2,18 +2,19 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import BarGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Bar/BarGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import barDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bar/barDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/axisInfoTextLabelsData';
+import getBarConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bar/barDefault';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/dataset1axisInfoLabels';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(barDefault('#axisInfoTextLabelsBargraph'));
-const dataset = utils.deepClone(data);
+
+const graphConfig = utils.deepClone(getBarConfig('#axisInfoTextLabelsBarGraph'));
+const dataset = [utils.deepClone(exampleData)];
 
 export default () => (
   <>
     <div id="tooltip" className="initial-tooltip" />
-    <BarGraph graphID="axisInfoTextLabelsBargraph" graphConfig={graphConfig} dataset={dataset} />
+    <BarGraph graphID="axisInfoTextLabelsBarGraph" graphConfig={graphConfig} dataset={dataset} />
   </>
 );