+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/BarGraph/BasicBarGraphs/StackedAxisInfoTextLabels.jsx
@@ -1,25 +1,26 @@
 import React from 'react';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import BarGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Bar/BarGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import barDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bar/barDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/stackedAxisInfoTextLabelsData';
+import getBarConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Bar/barDefault';
+import exampleData1 from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/stackedBarData1AxisInfo';
+import exampleData2 from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/stackedBarData2.js';
+import exampleData3 from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Bar/stackedBarData3.js';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(barDefault('#stackedAxisInfoTextLabelsBargraph'));
-graphConfig.axis.x.orientation = Carbon.helpers.AXES_ORIENTATION.X.TOP;
 
-const contentData = utils.deepClone(data);
-contentData[0].label.isDisabled = true;
-contentData[1].label.isDisabled = true;
-contentData[2].label.isDisabled = true;
+const graphConfig = utils.deepClone(getBarConfig('#stackedAxisInfoTextLabelsBarGraph'));
+const dataset = [
+  utils.deepClone(exampleData1),
+  utils.deepClone(exampleData2),
+  utils.deepClone(exampleData3),
+];
 
 export default () => (
   <>
     <div id="tooltip" className="initial-tooltip" />
-    <BarGraph graphID="stackedAxisInfoTextLabelsBargraph" graphConfig={graphConfig} dataset={contentData} />
+    <BarGraph graphID="stackedAxisInfoTextLabelsBarGraph" graphConfig={graphConfig} dataset={dataset} />
   </>
 );