+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/General/Legend/LineWithLegendOption.jsx
@@ -1,25 +1,27 @@
 import React from 'react';
-import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefault';
+import exampleData1 from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/General/dataset1';
+import exampleData2 from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/General/dataset2';
+import exampleData3 from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/General/dataset3';
 
 /*
 Please refer to the documentation below to see the graphConfig and data objects
 */
-const graphConfig = utils.deepClone(lineDefault('#lineWithLegendOption'));
 
-const contentData = utils.deepClone(data);
-contentData[0].showShapes = false;
-contentData[0].color = Carbon.helpers.COLORS.BLUE;
-contentData[0].style = { strokeDashArray: '2,2' };
+const graphConfig = utils.deepClone(getGraphConfig('#lineWithLegendOption'));
+const dataset = [
+  utils.deepClone(exampleData1),
+  utils.deepClone(exampleData2),
+  utils.deepClone(exampleData3),
+];
 
 export default () => (
   <>
     <div id="tooltip" className="initial-tooltip" />
-    <LineGraph graphID="lineWithLegendOption" graphConfig={graphConfig} dataset={contentData} />
+    <LineGraph graphID="lineWithLegendOption" graphConfig={graphConfig} dataset={dataset} />
   </>
 );