+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Grid/HorizontalGridHidden.test.jsx
@@ -2,9 +2,8 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultHGridHidden';
 
-const graphConfig = utils.deepClone(lineDefault('#horizontalGridHidden'));
-graphConfig.showHGrid = false;
+const graphConfig = utils.deepClone(getGraphConfig('#horizontalGridHidden'));
 
 export default () => <LineGraph graphID="horizontalGridHidden" graphConfig={graphConfig} />;