+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Grid/VerticalGridHidden.test.jsx
@@ -2,9 +2,8 @@ import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import LineGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Line/LineGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
-import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import getGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/generalDefaultVGridHidden';
 
-const graphConfig = utils.deepClone(lineDefault('#verticalGridHidden'));
-graphConfig.showVGrid = false;
+const graphConfig = utils.deepClone(getGraphConfig('#verticalGridHidden'));
 
 export default () => <LineGraph graphID="verticalGridHidden" graphConfig={graphConfig} />;